Added miui.app.* and com.qualcomm.qti.Performance

+++ src/ANXCamera/sources/com/android/camera/fragment/CtaNoticeFragment.java
@@ -80,7 +80,7 @@ public class CtaNoticeFragment extends DialogFragment {
     }
 
     public Dialog onCreateDialog(Bundle savedInstanceState) {
-        return new Builder(getActivity()).setTitle(R.string.user_notice_title).setMessage(buildUserNotice(getActivity(), R.string.user_notice_identify_summary_format)).setPositiveButton(R.string.user_agree, new OnClickListener() {
+        return new Builder(getActivity()).setTitle((int) R.string.user_notice_title).setMessage(buildUserNotice(getActivity(), R.string.user_notice_identify_summary_format)).setPositiveButton((int) R.string.user_agree, new OnClickListener() {
             public void onClick(DialogInterface dialog, int which) {
                 CTA.setCanConnectNetwork(((AlertDialog) CtaNoticeFragment.this.getDialog()).isChecked());
             }
